+++ include/material/Material.h
@@ -1,7 +1,7 @@
 #ifndef SIMPLE_RAY_TRACER_MATERIAL_H
 #define SIMPLE_RAY_TRACER_MATERIAL_H
 
-#include "util/math/Vector3d.h"
+#include "util/Vector3f.h"
 #include "texture/Texture.h"
 #include "surface/Facet.h"
 
@@ -15,15 +15,15 @@ public:
 
     explicit Material(Texture * texture): texture(texture) {}
 
-    virtual Vector3d getBRDF(const Vector3d & wi, const Vector3d & wo, const HitResult & hitResult) = 0;
+    virtual Vector3f getBRDF(const Vector3f & wi, const Vector3f & wo, const HitResult & hitResult) = 0;
 
-    virtual Vector3d sampleBRDF(const Vector3d & wo, float * pdf, const HitResult & hitResult) = 0;
+    virtual Vector3f sampleBRDF(const Vector3f & wo, float * pdf, const HitResult & hitResult) = 0;
 
-    virtual float getPdf(const Vector3d & wi, const Vector3d & wo, const HitResult & hitResult) = 0;
+    virtual float getPdf(const Vector3f & wi, const Vector3f & wo, const HitResult & hitResult) = 0;
 
     virtual bool isEmitting() = 0;
 
-    virtual Vector3d getEmitting(const Vector3d & wi, const HitResult & hitResult) = 0;
+    virtual Vector3f getEmitting(const Vector3f & wi, const HitResult & hitResult) = 0;
 
     virtual ~Material() = default;
