+++ include/material/TorranceSparrow.h
@@ -10,34 +10,34 @@ private:
 
     float alpha;
 
-    Vector3d f0;
+    Vector3f f0;
 
     Random randEp1;
 
     Random randEp2;
 
     // we'll use GGX here
-    float D(const Vector3d & wh, const Vector3d & n) const;
+    float D(const Vector3f & wh, const Vector3f & n) const;
 
     // we'll use smith height correlated masking-shadowing function
-    float G(const Vector3d & wi, const Vector3d & wo, const Vector3d & n) const;
+    float G(const Vector3f & wi, const Vector3f & wo, const Vector3f & n) const;
 
     // Fresnel item, we'll use schlick approximation
-    Vector3d F(const Vector3d & wo, const Vector3d & wh);
+    Vector3f F(const Vector3f & wo, const Vector3f & wh);
 
 public:
 
-    TorranceSparrow(const Vector3d & f0, float roughness, Texture * texture);
+    TorranceSparrow(const Vector3f & f0, float roughness, Texture * texture);
 
-    Vector3d getBRDF(const Vector3d & wi, const Vector3d & wo, const HitResult & hitResult) override;
+    Vector3f getBRDF(const Vector3f & wi, const Vector3f & wo, const HitResult & hitResult) override;
 
-    Vector3d sampleBRDF(const Vector3d & wo, float * pdf, const HitResult & hitResult) override;
+    Vector3f sampleBRDF(const Vector3f & wo, float * pdf, const HitResult & hitResult) override;
 
-    float getPdf(const Vector3d & wi, const Vector3d & wo, const HitResult & hitResult) override;
+    float getPdf(const Vector3f & wi, const Vector3f & wo, const HitResult & hitResult) override;
 
     bool isEmitting() override;
 
-    Vector3d getEmitting(const Vector3d & wi, const HitResult & hitResult) override;
+    Vector3f getEmitting(const Vector3f & wi, const HitResult & hitResult) override;
 
     ~TorranceSparrow() override = default;
