+++ include/render/MISampler.h
@@ -18,13 +18,13 @@ private:
 
     static void render0(Camera * camera, Scene * scene, ImageBuffer * im, int spp);
 
-    static Vector3d trace0(const Ray & ray, Scene * scene);
+    static Vector3f trace0(const Ray & ray, Scene * scene);
 
-    static Vector3d shade0(const Vector3d & wo, Scene * scene, const HitResult & hitResult);
+    static Vector3f shade0(const Vector3f & wo, Scene * scene, const HitResult & hitResult);
 
-    static Vector3d sampleLight0(const Vector3d & wo, Scene * scene, const HitResult & hitResult);
+    static Vector3f sampleLight0(const Vector3f & wo, Scene * scene, const HitResult & hitResult);
 
-    static Vector3d sampleBRDF0(const Vector3d & wo, Scene * scene, const HitResult & hitResult);
+    static Vector3f sampleBRDF0(const Vector3f & wo, Scene * scene, const HitResult & hitResult);
 
 public:
